Finished Sending Emails

+++ cms/admin/includes/view_all_users.php
@@ -17,15 +17,15 @@
     $query = "SELECT * FROM users";
     $select_users = mysqli_query($connection, $query);
 
-    while($row = mysqli_fetch_assoc($select_users)) {
-      $user_id = $row['user_id'];
-      $username = $row['username'];
-      $user_password = $row['user_password'];
-      $user_firstname = $row['user_firstname'];
-      $user_lastname = $row['user_lastname'];
-      $user_email = $row['user_email'];
-      $user_image = $row['user_image'];
-      $user_role = $row['user_role'];
+    while($row        = mysqli_fetch_assoc($select_users)) {
+      $user_id        = escape($row['user_id']);
+      $username       = escape($row['username']);
+      $user_password  = escape($row['user_password']);
+      $user_firstname = escape($row['user_firstname']);
+      $user_lastname  = escape($row['user_lastname']);
+      $user_email     = escape($row['user_email']);
+      $user_image     = escape($row['user_image']);
+      $user_role      = escape($row['user_role']);
       
       echo "<tr>";
       
@@ -89,24 +89,30 @@
 <?php 
 
 if(isset($_GET['change_to_admin'])) {
-  $the_user_id = $_GET['change_to_admin'];
+  $the_user_id = escape($_GET['change_to_admin']);
   $query = "UPDATE users SET user_role = 'admin' WHERE user_id = {$the_user_id} ";
   $change_to_admin_query = mysqli_query($connection, $query);
   header("Location: users.php");
 }
 
 if(isset($_GET['change_to_sub'])) {
-  $the_user_id = $_GET['change_to_sub'];
+  $the_user_id = escape($_GET['change_to_sub']);
   $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = {$the_user_id} ";
   $change_to_sub_query = mysqli_query($connection, $query);
   header("Location: users.php");
 }
 
 if(isset($_GET['delete'])) {
-  $the_user_id = $_GET['delete'];
-  $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-  $delete_query = mysqli_query($connection, $query);
-  header("Location: users.php");
+  if(isset($_SESSION['user_role'])) {
+    if($_SESSION['user_role'] == 'admin') {
+      $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+      $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+      $delete_query = mysqli_query($connection, $query);
+      header("Location: users.php");
+    }
+  }
+
+
 }
